Moved some math macros (sqr, lerp, clamp) to libslic3r.h
Added UNUSED macro to libslic3r.h, used it to reduce some compile warnings.

Split the Int128 class from Clipper library to a separate file,
extended Int128 with intrinsic types wherever possible for performance,
added new geometric predicates.

Added a draft of new FillRectilinear3, which should reduce overfill near the perimeters in the future.

+++ xs/src/libslic3r/Geometry.hpp
@@ -49,7 +49,7 @@ static inline bool is_ccw(const Polygon &poly)
 
     // 1) Find the lowest lexicographical point.
     int     imin = 0;
-    for (size_t i = 1; i < poly.points.size(); ++ i) {
+    for (int i = 1; i < poly.points.size(); ++ i) {
         const Point &pmin = poly.points[imin];
         const Point &p    = poly.points[i];
         if (p.x < pmin.x || (p.x == pmin.x && p.y < pmin.y))
